Refactor(Console, Others): 수익률 double로 변경.

+++ src/main/java/lotto/console/Output.java
@@ -53,7 +53,7 @@ public class Output {
                 count);
     }
 
-    public void writeYield(float yield) {
+    public void writeYield(double yield) {
         writeLine("총 수익률은 %s입니다.", formatPercent(yield));
     }
